+++ src/core/processors/normalization.ts
@@ -1,53 +1,36 @@
-/** Function to process html string for normalization*/
-export type UIPNormalizationProcessor = (input: string) => string;
+import {UIPPreprocessorService} from './preprocessor';
 
-export class UIPHtmlNormalizationService {
-  /** Processor storage */
-  protected static processors: Record<string, UIPNormalizationProcessor> = {};
+/**
+ * Normalization processors store for JS content.
+ * Normalization transformation are used to normalize content before displaying or processing.
+ */
+export const UIPJSNormalizationPreprocessors = new UIPPreprocessorService();
+/**
+ * Normalization processors store for HTML content.
+ * Normalization transformation are used to normalize content before displaying or processing.
+ */
+export const UIPHTMLNormalizationPreprocessors = new UIPPreprocessorService();
 
-  /** Add processor function to normalization process */
-  public static addProcessor(
-    name: string,
-    processor: UIPNormalizationProcessor
-  ): void {
-    UIPHtmlNormalizationService.processors[name] = processor;
-  }
 
-  /** Normalizes passes html string by running all registered processors in chain */
-  public static normalize(html: string): string {
-    return Object.keys(UIPHtmlNormalizationService.processors)
-      .map((name) => UIPHtmlNormalizationService.processors[name])
-      .filter((processor) => typeof processor === 'function')
-      .reduce((input, processor) => processor(input), html);
-  }
+/** Removes extra indents to beautify content alignment */
+export function removeIndent(input: string): string {
+  // Get all indents from text
+  const indents = input.match(/^[^\S\n\r]*(?=\S)/gm);
+  // No processing if no indent on the first line or input is empty
+  if (!indents || !indents[0].length) return input;
+  // Sort indents by length
+  indents.sort((a, b) => a.length - b.length);
+  // No processing if minimal indent is 0
+  if (!indents[0].length) return input;
+  // Remove indents from text
+  return input.replace(RegExp('^' + indents[0], 'gm'), '');
 }
+UIPJSNormalizationPreprocessors.add('remove-leading-indent', removeIndent);
+UIPHTMLNormalizationPreprocessors.add('remove-leading-indent', removeIndent);
 
-/** Removes extra indents */
-UIPHtmlNormalizationService.addProcessor(
-  'remove-indent',
-  (input: string): string => {
-    // Get all indents from text
-    const indents = input.match(/^[^\S\n\r]*(?=\S)/gm);
-    // No processing if no indent on the first line or input is empty
-    if (!indents || !indents[0].length) return input;
-    // Sort indents by length
-    indents.sort((a, b) => a.length - b.length);
-    // No processing if minimal indent is 0
-    if (!indents[0].length) return input;
-    // Remove indents from text
-    return input.replace(RegExp('^' + indents[0], 'gm'), '');
-  }
-);
-/** Removes extra spaces */
-// TODO: handle case with inline script with literal double spaces
-UIPHtmlNormalizationService.addProcessor('remove-trailing', (input: string) =>
-  input.replace(/\s*?$/gm, '')
-);
-/** Remove beginning spaces */
-UIPHtmlNormalizationService.addProcessor('left-trim', (input: string) =>
-  input.replace(/^\s+/, '')
-);
-/** Remove ending spaces */
-UIPHtmlNormalizationService.addProcessor('right-trim', (input: string) =>
-  input.replace(/\s+$/, '')
-);
+/** Trim content */
+UIPJSNormalizationPreprocessors.add('trim', (content: string) => content.trim());
+UIPHTMLNormalizationPreprocessors.add('trim', (content: string) => content.trim());
+
+/** Removes extra spaces inside the content. Applicable for HTML only */
+UIPHTMLNormalizationPreprocessors.addRegexReplacer('remove-trailing', /\s*?$/gm, '');